requirements.txt
Pillow
torch
torchvision
transformers
sentencepiece
opencv-python
accelerate
tiktoken
PyMuPDF
python-docx
python-pptx